requirements.txt
google-genai>=1.0.0
gradio>=4.0.0
Pillow>=10.0.0
loguru>=0.7.0
python-dotenv>=1.0.0
tabulate>=0.9.0
ultralytics>=8.3.0
